requirements.txt
loguru==0.7.2
python-dotenv==0.21.0
boto3==1.26.47
botocore==1.29.47
awscli-local==0.20
psycopg2==2.9.9